guitty: better coalescing in progressbar rendering
Problem: ttyrec files can often contain long pauses (e.g. due to AFK).
We want the video player progress bar to show where these pauses are in
the file.

Solution: draw a light gray "tick" for every frame based on its position
in the recording. Then long pauses will be visible in the progress bar
as dark regions.

Problem: if we draw a rectangle for every frame in the video, we will be
sending potentially thousands of draw calls to the renderer, which is
unnecessary as most of these rectangles will be very narrow and overlap
each other.

Solution: "coalesce" these rectangles by determining when they would
overlap and drawing a larger rectangle instead.

Problem: find a data structure which supports querying for all
subsequences of frames we've loaded so far, where the difference in
timestamps between each pair of adjacent frames in a subsequence is
sufficiently small - small enough that the ticks would overlap in the
progress bar, so:

  (Δt / T) < (w / W)

  where Δt is the timestamp difference,
        T  is the time between the first and last frame,
        w  is the width of a tick,
        W  is the width of the progress bar

These queries need to be efficient: specifically, they should take
O(# of subsequences) time, not O(# of frames).

==========

The previous way we solved this problem is to store all the subsequences
in an array based on how they were coalesced on the previous render.
Then, whenever T * (w / W) increases (usually because T increases as we
load more frames), we could rescan the array, potentially coalescing
more subsequences. This uses only O(# subsequences) memory as a bonus.

The only problem with this approach is that it's lossy. We permanently
remove distinctions between subsequences when T / W increases, but this
means we have to rescan the entire array when the user expands the
window (increasing W).

The new approach solves this by organizing frames into a binary tree.
The branches of the tree are stored in `progress_bar.State`, and the
leaves are just stored as indexes into the frame array.

There is one branch for every gap between two frames, and the tree is
balanced such that the root of any subtree represents a gap no smaller
than any of its children. Conceptually, adding an element to the end
looks like this:

    type a
    type separation = int
    type node = Leaf   of a
              | Branch of separation * node * node

    let rec append (s : separation) (x : a) node =
      match node with
      | Branch (s', lhs, rhs) when s < s' ->
             Branch (s', lhs, append s x rhs)
      | _ -> Branch (s, node, Leaf x)

We can use this to do the query in O(# subsequences) time, because
once we find a branch with a sufficiently small gap, we do not need to
check its children. Instead, we know that the entirety of the branch
represents a subsequence. However, we never actually discard the
children.

This means we use O(# frames) memory, however:
- We avoid re-scanning the entire subsequences array when frames are
  added or the window is made smaller;
- We avoid re-scanning the frame array when the window is widened.
- It's only 16 bytes per frame, which is less than we already store per
  frame in the frame array.
- There's nothing stopping us from doing the same optimization of
  trimming the tree in the future, if we decide that a group of frames
  is so close together that there's no way we would ever want to draw
  them separately anymore.